Move dependency analysis framework files into separate folder/namespace
