+++ ILCompiler/Compiler/DependencyAnalysis/ConstructedEETypeNode.cs
@@ -1,5 +1,6 @@
 ﻿using dnlib.DotNet;
 using ILCompiler.Common.TypeSystem.Common;
+using ILCompiler.Compiler.DependencyAnalysisFramework;
 
 namespace ILCompiler.Compiler.DependencyAnalysis
 {
