+++ ILCompiler/Compiler/DependencyAnalysisFramework/ConditionalDependency.cs
@@ -1,4 +1,4 @@
-﻿namespace ILCompiler.Compiler.DependencyAnalysis
+﻿namespace ILCompiler.Compiler.DependencyAnalysisFramework
 {
     public record ConditionalDependency
     {
